# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ConstraintCreate": {
        "description": "Schema for creating a constraint.",
        "properties": {
          "key": {
            "description": "Constraint key/identifier",
            "title": "Key",
            "type": "string"
          },
          "org_id": {
            "description": "Organization ID",
            "title": "Org Id",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Constraint parameters",
            "title": "Params"
          },
          "predicate": {
            "description": "Constraint predicate/rule",
            "title": "Predicate",
            "type": "string"
          },
          "type": {
            "description": "Constraint type: hard or soft",
            "title": "Type",
            "type": "string"
          },
          "weight": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Weight for soft constraints",
            "title": "Weight"
          }
        },
        "required": [
          "key",
          "type",
          "predicate",
          "org_id"
        ],
        "title": "ConstraintCreate",
        "type": "object"
      },
      "ConstraintResponse": {
        "description": "Schema for constraint response.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "key": {
            "description": "Constraint key/identifier",
            "title": "Key",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Constraint parameters",
            "title": "Params"
          },
          "predicate": {
            "description": "Constraint predicate/rule",
            "title": "Predicate",
            "type": "string"
          },
          "type": {
            "description": "Constraint type: hard or soft",
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "weight": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Weight for soft constraints",
            "title": "Weight"
          }
        },
        "required": [
          "key",
          "type",
          "predicate",
          "id",
          "org_id",
          "created_at",
          "updated_at"
        ],
        "title": "ConstraintResponse",
        "type": "object"
      },
      "ConstraintUpdate": {
        "description": "Schema for updating a constraint.",
        "properties": {
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "predicate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicate"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "weight": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          }
        },
        "title": "ConstraintUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListResponse_ConstraintResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConstraintResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ListResponse[ConstraintResponse]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered volunteer scheduling and sign-up management",
    "title": "SignUpFlow API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/constraints/": {
      "get": {
        "description": "List constraints with optional filters.",
        "operationId": "listConstraints",
        "parameters": [
          {
            "description": "Filter by organization ID",
            "in": "query",
            "name": "org_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by organization ID",
              "title": "Org Id"
            }
          },
          {
            "description": "Filter by type (hard/soft)",
            "in": "query",
            "name": "constraint_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by type (hard/soft)",
              "title": "Constraint Type"
            }
          },
          {
            "description": "Page size, max 200",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size, max 200",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of rows to skip",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of rows to skip",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_ConstraintResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Constraints",
        "tags": [
          "constraints"
        ]
      },
      "post": {
        "description": "Create a new constraint.",
        "operationId": "createConstraint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstraintCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstraintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Constraint",
        "tags": [
          "constraints"
        ]
      }
    },
    "/api/v1/constraints/{constraint_id}": {
      "delete": {
        "description": "Delete constraint.",
        "operationId": "deleteConstraint",
        "parameters": [
          {
            "in": "path",
            "name": "constraint_id",
            "required": true,
            "schema": {
              "title": "Constraint Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Constraint",
        "tags": [
          "constraints"
        ]
      },
      "get": {
        "description": "Get constraint by ID.",
        "operationId": "getConstraint",
        "parameters": [
          {
            "in": "path",
            "name": "constraint_id",
            "required": true,
            "schema": {
              "title": "Constraint Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstraintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Constraint",
        "tags": [
          "constraints"
        ]
      },
      "put": {
        "description": "Update constraint.",
        "operationId": "updateConstraint",
        "parameters": [
          {
            "in": "path",
            "name": "constraint_id",
            "required": true,
            "schema": {
              "title": "Constraint Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstraintUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstraintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Constraint",
        "tags": [
          "constraints"
        ]
      }
    },
...
  }
}